# openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/v1/ledger/orders/{orderId}/total" : {
      "get" : {
        "tags" : [ "ledger-controller" ],
        "operationId" : "getOrderLedgerAmount",
        "parameters" : [ {
          "name" : "orderId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "number"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger/cash/drivers/{driverId}" : {
      "get" : {
        "tags" : [ "payee-cash-controller" ],
        "operationId" : "getCashByDriver",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoCashRemittanceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger/cash/drivers/{driverId}/summary" : {
      "get" : {
        "tags" : [ "payee-cash-controller" ],
        "operationId" : "getCashSummary",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CashSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/ledger/accounts/{ownerType}/{ownerId}" : {
      "get" : {
        "tags" : [ "ledger-controller" ],
        "operationId" : "getAccount",
        "parameters" : [ {
          "name" : "ownerType",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "enum" : [ "GATEWAY_RECEIVABLE", "CASH_RECEIVABLE", "BANK", "PLATFORM_CLEARING", "PLATFORM_REVENUE", "TAX_PAYABLE", "PAYOUT_IN_TRANSIT", "RESTAURANT_PAYABLE", "DRIVER_PAYABLE", "CUSTOMER_CREDIT", "ADVERTISER_PREPAID" ]
          }
        }, {
          "name" : "ownerId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LedgerAccount"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/ledger/cash/drivers/{driverId}" : {
      "get" : {
        "tags" : [ "internal-cash-controller" ],
        "operationId" : "getCashByDriver_1",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseDtoCashRemittanceDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/ledger/cash/drivers/{driverId}/summary" : {
      "get" : {
        "tags" : [ "internal-cash-controller" ],
        "operationId" : "getCashSummary_1",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CashSummaryDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/internal/admin/cash/drivers/{driverId}" : {
      "get" : {
        "tags" : [ "cash-controller" ],
        "operationId" : "getCashByDriver_2",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageCashRemittance"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/internal/admin/cash/drivers/{driverId}/summary" : {
      "get" : {
        "tags" : [ "cash-controller" ],
        "operationId" : "getCashSummary_2",
        "parameters" : [ {
          "name" : "driverId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CashSummaryDto"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "CashRemittance" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "driverId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "reference" : {
            "type" : "string"
          },
          "recordedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ledgerTransactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PageableObject" : {
        "required" : [ "offset", "pageNumber", "pageSize", "paged", "unpaged" ],
        "type" : "object",
        "properties" : {
          "offset" : {
            "type" : "integer",
            "format" : "int64"
          },
          "pageSize" : {
            "type" : "integer",
            "format" : "int32"
          },
          "paged" : {
            "type" : "boolean"
          },
          "pageNumber" : {
            "type" : "integer",
            "format" : "int32"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "unpaged" : {
            "type" : "boolean"
          }
        }
      },
      "SortObject" : {
        "required" : [ "empty", "sorted", "unsorted" ],
        "type" : "object",
        "properties" : {
          "empty" : {
            "type" : "boolean"
          },
          "sorted" : {
            "type" : "boolean"
          },
          "unsorted" : {
            "type" : "boolean"
          }
        }
      },
...
      "CashRemittanceDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "driverId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "amount" : {
            "type" : "number"
          },
          "reference" : {
            "type" : "string"
          },
          "recordedBy" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ledgerTransactionId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "PageResponseDtoCashRemittanceDto" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CashRemittanceDto"
            }
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "last" : {
            "type" : "boolean"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "first" : {
            "type" : "boolean"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      },
      "CashSummaryDto" : {
        "type" : "object",
        "properties" : {
          "driverId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "cashCollected" : {
            "type" : "number"
          },
          "cashRemitted" : {
            "type" : "number"
          },
          "cashInHand" : {
            "type" : "number"
          }
        }
      },
      "LedgerAccount" : {
        "required" : [ "balance", "id", "lockVersion", "ownerId", "ownerType" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "ownerType" : {
            "type" : "string",
            "enum" : [ "GATEWAY_RECEIVABLE", "CASH_RECEIVABLE", "BANK", "PLATFORM_CLEARING", "PLATFORM_REVENUE", "TAX_PAYABLE", "PAYOUT_IN_TRANSIT", "RESTAURANT_PAYABLE", "DRIVER_PAYABLE", "CUSTOMER_CREDIT", "ADVERTISER_PREPAID" ]
          },
          "ownerId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "kind" : {
            "type" : "string",
            "enum" : [ "EXTERNAL", "INTERNAL", "PAYABLE", "PREPAID" ]
          },
          "balance" : {
            "type" : "number"
          },
          "currency" : {
            "type" : "string"
          },
          "lockVersion" : {
            "type" : "integer",
            "format" : "int32"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "PageCashRemittance" : {
        "required" : [ "content", "empty", "first", "last", "number", "numberOfElements", "size", "totalElements", "totalPages" ],
        "type" : "object",
        "properties" : {
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "numberOfElements" : {
            "type" : "integer",
            "format" : "int32"
          },
          "number" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/CashRemittance"
            }
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          },
          "pageable" : {
            "$ref" : "#/components/schemas/PageableObject"
          },
          "sort" : {
            "$ref" : "#/components/schemas/SortObject"
          },
          "empty" : {
            "type" : "boolean"
          }
        }
      }
    }
  }
}